Update khandler
The original k handler indexed into the file to extract data assuming a specified spacing between node definitions and with a space delimiter. However the generation of .k files is not generic and therefore one with a different spacing between node positions will not work indexing numbers. The updated version uses a regex expression to extract data and is independent of the spacing or comma/space delimiter type.